Configure Document Scanning from server and save camera metadata (#1007)
- Configures the `DocumentScanner` from values sent by the server response
- Saves camera metadata to server when saving the image

+++ StripeIdentity/StripeIdentity/Source/API Bindings/Models/VerificationPage/VerificationPageStaticContentDocumentCapturePage.swift
@@ -23,6 +23,6 @@ struct VerificationPageStaticContentDocumentCapturePage: Decodable, Equatable {
     let lowResImageMaxDimension: Int
     let models: VerificationPageStaticContentDocumentCaptureModels
     let motionBlurMinDuration: Int
-    let motionBlurMinIou: CGFloat
+    let motionBlurMinIou: Decimal
     let requireLiveCapture: Bool
 }
